% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpalette.R
\name{iddu}
\alias{iddu}
\title{IDDU palette}
\usage{
iddu(n = NULL)
}
\arguments{
\item{n}{\code{numeric}. How many colours?}
}
\value{
\code{character} of length \code{n}
}
\description{
An alias for \code{idpalette(p = "iddu", n)}
}
\examples{
iddu(5)
}
