% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem.R
\name{idem}
\alias{idem}
\title{idem palette}
\usage{
idem(n)
}
\arguments{
\item{n}{\code{numeric}. How many colours?}
}
\value{
\code{character} of length \code{n}
}
\description{
An alias for \code{idpalette(p = "idem", n)}
}
\examples{
idem(7)
}
