% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpalette.R
\name{thekids}
\alias{thekids}
\title{The Kids palette}
\usage{
thekids(n = NULL, rev = FALSE)
}
\arguments{
\item{n}{\code{numeric}. How many colours?}

\item{rev}{\code{logical} Reverse the colour order?}
}
\value{
\code{idpalette} class object of hex colours length \code{n}.
The \code{idpalette} class has a print method that will plot the colours in the
object, but it is underneath that a \code{character} string of length \code{n} and can
be otherwise treated as such.
}
\description{
An alias for \code{idpalette(p = "thekids", n, rev)}
}
\examples{
thekids(8)
}
