% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpalette.R
\name{acefa}
\alias{acefa}
\title{ACEFA palette}
\usage{
acefa(n = NULL)
}
\arguments{
\item{n}{\code{numeric}. How many colours?}
}
\value{
\code{character} of length \code{n}
}
\description{
An alias for \code{idpalette(p = "acefa", n)}
}
\examples{
acefa(5)
}
