% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpal.R
\name{idpal}
\alias{idpal}
\title{Base colour palettes for IDEM, IDDU, and ACEFA}
\usage{
idpal(p = c("idem", "iddu", "acefa", "acefa_pink", "idem_official"))
}
\arguments{
\item{p}{\code{character}. Which palette do you want?}
}
\value{
\code{character} of hex values
}
\description{
Base colour palettes for IDEM, IDDU, and ACEFA
}
\details{
\code{"idem"}, \code{"iddu"}, and \code{"acefa"} return palettes for all colours in logos.
\code{"..._official"} palettes return only the four-colour palettes per the style-
guide from the designer.

NB: These 'non-official' colours are estimated by a colour dropper thing
Gerry found on the internet and may or may not be exacty correct.
}
\examples{

idpal("idem")
}
