% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpalette-package.R
\docType{package}
\name{idpalette-package}
\alias{idpalette-package}
\title{idpalette: Palettes for IDEM, IDDU, ACEFA, and The Kids}
\description{
See: package title.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://idem-lab.github.io/idpalette/}
  \item \url{https://github.com/idem-lab/idpalette}
  \item Report bugs at \url{https://github.com/idem-lab/idpalette/issues}
}

}
\author{
\strong{Maintainer}: Gerry Ryan \email{id@em.com} (\href{https://orcid.org/0000-0003-0183-7630}{ORCID})

}
\keyword{internal}
