% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpalette-package.R
\docType{package}
\name{idpalette-package}
\alias{idpalette-package}
\title{idpalette: Some palettes for the IDEM, IDDU, and ACEFA research groups}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://idem-lab.github.io/idpalette/}
}

}
\author{
\strong{Maintainer}: Gerry Ryan \email{id@em.com} (\href{https://orcid.org/0000-0003-0183-7630}{ORCID})

}
\keyword{internal}
