% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpalette.R
\name{idpalette}
\alias{idpalette}
\title{ID Palette}
\usage{
idpalette(p, n = NULL)
}
\arguments{
\item{p}{\code{character}. Which palette? See \code{?idpal} for details.}

\item{n}{\code{numeric}. How many colours}
}
\value{
\code{character} of hex colours length \code{n}
}
\description{
Create colour palettes based on IDEM, IDDU, and ACEFA colours
}
\examples{

idpalette("idem", 20)
}
