% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_id_continuous.R
\name{scale_id_continuous}
\alias{scale_id_continuous}
\title{scale ID continuous}
\usage{
scale_id_continuous(
  cols,
  aesthetics = c("fill", "colour"),
  na.value = "transparent",
  guide = "colourbar"
)
}
\arguments{
\item{cols}{\code{character}. A vector of colours}

\item{aesthetics}{Scale aesthetic.}

\item{na.value}{Garn - whaddaya want in the \code{NA}s?}

\item{guide}{A function used to create a guide or its name. See \code{\link[ggplot2:guides]{ggplot2::guides()}} for more information}
}
\value{
A \code{ggproto} object of class \code{Scale} and \code{ScaleContinuous}
}
\description{
Convenience wrapper to \code{ggplot2::continuous_scale} to allow use of
\code{idpalette} or any arbitrary set of colours to be used as a \code{ggplot2} fill
or colour scale
}
\examples{

library(ggplot2)

ggplot(mtcars) +
  geom_point(
    aes(
      x = disp,
      y = hp,
      colour = qsec
    ),
    size = 5
  ) +
  scale_id_continuous(
    cols = iddu(),
    aesthetics = "colour"
  )

}
